+++ tests/testthat/test-interface.R
@@ -1,11 +1,16 @@
 S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
 X <- data.frame(x = c("a", "b"), y = c(100, 10))
 shp <- shapviz(S, X, baseline = 4)
+mshp <- c(shp = shp, shp2 = shp + shp)
 
 test_that("get_* functions work", {
   expect_equal(4, get_baseline(shp))
   expect_equal(S, get_shap_values(shp))
   expect_equal(X, get_feature_values(shp))
+
+  expect_equal(4, get_baseline(mshp)[[1L]])
+  expect_equal(S, get_shap_values(mshp)[[1L]])
+  expect_equal(X, get_feature_values(mshp)[[1L]])
 })
 
 test_that("dim, nrow, ncol, colnames work", {
@@ -18,11 +23,11 @@ test_that("dim, nrow, ncol, colnames work", {
 test_that("subsetting works", {
   expect_equal(dim(shp[, "x"]$S), c(2L, 1L))
   expect_equal(dim(shp[, "x"]$X), c(2L, 1L))
-  expect_equal(dim(shp[1, "x"]$S), c(1L, 1L))
-  expect_equal(dim(shp[1, "x"]$X), c(1L, 1L))
-  expect_equal(dim(shp[1, ]$S), c(1L, 2L))
-  expect_equal(dim(shp[1, ]$X), c(1L, 2L))
-  expect_equal(get_baseline(shp[1, ]), get_baseline(shp))
+  expect_equal(dim(shp[1L, "x"]$S), c(1L, 1L))
+  expect_equal(dim(shp[1L, "x"]$X), c(1L, 1L))
+  expect_equal(dim(shp[1L, ]$S), c(1L, 2L))
+  expect_equal(dim(shp[1L, ]$X), c(1L, 2L))
+  expect_equal(get_baseline(shp[1L, ]), get_baseline(shp))
 })
 
 test_that("concatenating with + works", {
@@ -43,6 +48,11 @@ test_that("concatenating with rbind works", {
 
 test_that("print() gives no error", {
   capture_output(expect_no_error(print(shp)))
+  capture_output(expect_no_error(print(mshp)))
+})
+
+test_that("summary() gives no error", {
+  capture_output(expect_no_error(summary(shp)))
 })
 
 test_that("shapviz() gives error for default method", {
@@ -141,6 +151,7 @@ S_inter <- array(
   dimnames = list(NULL, c("x", "y"), c("x", "y"))
 )
 shp_inter <- shapviz(S, X = X, baseline = 4, S_inter = S_inter)
+mshp_inter <- c(shp1 = shp_inter, shp2 = shp_inter + shp_inter)
 
 test_that("get_shap_interactions, +, rbind works for interactions", {
   expect_equal(S_inter, get_shap_interactions(shp_inter))
@@ -151,11 +162,18 @@ test_that("get_shap_interactions, +, rbind works for interactions", {
   )
 })
 
-test_that("print() gives no error (with interactions)", {
+test_that("print() and summary() gives no error (with interactions)", {
   capture_output(expect_no_error(print(shp_inter)))
+  capture_output(expect_no_error(print(mshp_inter)))
+  capture_output(expect_no_error(summary(shp_inter)))
 })
 
-# Multiclass
+test_that("mshapviz object contains original shapviz objects", {
+  expect_equal(mshp_inter[[1L]], shp_inter)
+  expect_equal(mshp_inter[[2L]][1:nrow(shp_inter)], shp_inter)
+})
+
+# Multiclass with XGBoost
 X_pred <- data.matrix(iris[, -5L])
 dtrain <- xgboost::xgb.DMatrix(X_pred, label = as.integer(iris[, 5L]) - 1L)
 fit <- xgboost::xgb.train(
@@ -165,14 +183,29 @@ fit <- xgboost::xgb.train(
   objective="multi:softprob",
   num_class=3L
 )
+shp3 <- shapviz(fit, X_pred = X_pred, which_class = 3L, interactions = TRUE)
+mshp <- shapviz(fit, X_pred = X_pred, interactions = TRUE)
 
-test_that("initialization of multiclass xgb model fails without 'which_class'", {
-  expect_error(shapviz(fit, X_pred = X_pred))
+test_that("is.shapviz() and is.mshapviz() functions work", {
+  expect_true(is.shapviz(shp3))
+  expect_true(is.mshapviz(mshp))
+  expect_false(is.shapviz(mshp))
+  expect_false(is.mshapviz(shp3))
 })
 
-test_that("initialization of multiclass xgb model succeeds with 'which_class'", {
-  expect_true(
-    is.shapviz(shapviz(fit, X_pred = X_pred, which_class = 3L, interactions = TRUE))
-  )
+test_that("shapviz on class 3 equals mshapviz[[3]] for classification", {
+  expect_equal(mshp[[3L]], shp3)
+})
+
+test_that("combining shapviz on classes 1, 2, 3 equal mshapviz", {
+  shp1 <- shapviz(fit, X_pred = X_pred, which_class = 1L, interactions = TRUE)
+  shp2 <- shapviz(fit, X_pred = X_pred, which_class = 2L, interactions = TRUE)
+  expect_equal(mshp, c(Class_1 = shp1, Class_2 = shp2, Class_3 = shp3))
+  expect_equal(mshp, mshapviz(list(Class_1 = shp1, Class_2 = shp2, Class_3 = shp3)))
+})
+
+test_that("combining non-shapviz objects fails", {
+  expect_error(c(shp3, 1))
+  expect_error(mshapviz(1, 2))
 })